refactor(backup): extract closure + row-transforms into a focused module
Move the referential-closure checks and row-remap transforms out of the
1.5k-line company-backup.ts into company-backup.transforms.ts:
findDanglingReferences, assertReferentiallyClosed, loadSubstrateIds,
buildRowTransforms, USER_REF_TABLES, DanglingRef/RowTransform.

company-backup.ts keeps catalog introspection, serialization, storage and
scope-query helpers. The new module imports one-directionally from it (no
cycle); RETAINED_REF_TABLES stays in core (findExportScopeViolations uses it)
and is imported back. Consumers (restore/import/closure test) repointed.

Behavior-preserving — 28 closure tests + typecheck unchanged. Net: company-backup.ts
drops 1368 -> 1210 lines (below its pre-change size).
